+++ rnalysis/__init__.py
@@ -4,7 +4,7 @@
 import warnings
 
 __all__ = ['general', 'filtering', 'enrichment', '__attr_file_key__', '__biotype_file_key__', '__font_key__',
-           '__font_size_key__', '__stylesheet_key__']
+           '__font_size_key__', '__stylesheet_key__', '__show_tutorial_key__','__databases_key__']
 __name__ = "rnalysis"
 __author__ = "Guy Teichman"
 __email__ = "guyteichman@gmail.com"
@@ -16,6 +16,7 @@ __font_key__ = "gui_font"
 __font_size_key__ = "gui_font_size"
 __stylesheet_key__ = "gui_stylesheet"
 __show_tutorial_key__ = "show_tutorial_on_startup"
+__databases_key__ = "databases"
 
 
 def _simple_warning_format(msg, *args, **kwargs):