+++ rnalysis/utils/settings.py
@@ -5,11 +5,11 @@ from typing import Union, Any
 import yaml
 
 from rnalysis import __attr_file_key__, __biotype_file_key__, __font_key__, __font_size_key__, __stylesheet_key__, \
-    __show_tutorial_key__
+    __show_tutorial_key__, __databases_key__
 from rnalysis.utils import io
 
 DEFAULT_VALUES = {__font_key__: 'Times New Roman', __font_size_key__: 10, __stylesheet_key__: 'base',
-                  __show_tutorial_key__: True}
+                  __show_tutorial_key__: True, __databases_key__: ['Google', 'NCBI Genes', 'UniProtKB']}
 
 
 def get_settings_file_path():
@@ -141,15 +141,21 @@ def get_gui_settings():
     font = read_value_from_settings(__font_key__)
     font_size = int(read_value_from_settings(__font_size_key__))
     stylesheet = read_value_from_settings(__stylesheet_key__)
+    databases = read_value_from_settings(__databases_key__)
     show_tutorial = read_value_from_settings(__show_tutorial_key__)
-    return font, font_size, stylesheet, show_tutorial
+    return font, font_size, stylesheet, databases, show_tutorial
 
 
-def set_gui_settings(font: str, font_size: int, stylesheet: str, show_tutorial: bool):
+def set_gui_settings(font: str, font_size: int, stylesheet: str, databases, show_tutorial: bool):
     update_settings_file(font, __font_key__)
     update_settings_file(str(font_size), __font_size_key__)
     update_settings_file(stylesheet, __stylesheet_key__)
+    update_settings_file(databases, __databases_key__)
     update_settings_file(show_tutorial, __show_tutorial_key__)
+
+
+def get_databases_settings():
+    return read_value_from_settings(__databases_key__)
 
 
 def get_show_tutorial_settings():